Made the JoyCon start it's own separate thread and take action requests. Sometimes you'll call methods within the JoyCon class on another thread that calls this.wait(), making the calling thread wait instead of the JoyCon. This way, the JoyCon will wait when it's supposed to. The JoyCon test now works however, upon sending output and then shutting down, it sends one more output report. During the Rumble test, all was working well, until I shutdown the program which made the JoyCon rumble again. Will fix this.

+++ src/main/java/couch/joycouch/analog/AnalogStick.java
@@ -1,18 +1,18 @@
 package couch.joycouch.analog;
 
 public class AnalogStick {
-    private int horizontal, vertical;
+    private float horizontal, vertical;
 
-    public AnalogStick(int horizontal, int vertical){
+    public AnalogStick(float horizontal, float vertical){
         this.horizontal = horizontal;
         this.vertical = vertical;
     }
 
-    public int getHorizontal(){
+    public float getHorizontal(){
         return this.horizontal;
     }
 
-    public int getVertical(){
+    public float getVertical(){
         return this.vertical;
     }
 }
