+++ src/main/java/couch/joycouch/analog/AnalogStickStatus.java
@@ -2,13 +2,13 @@ package couch.joycouch.analog;
 
 public class AnalogStickStatus {
 
-    private int horizontal, vertical;
+    private float horizontal, vertical;
 
     public AnalogStickStatus(
             AnalogStickCalibrator calibrator,
             byte data1, byte data2, byte data3
     ){
-        int[] leftPos = calibrator.getStickPos(data1, data2, data3);
+        float[] leftPos = calibrator.getStickPos(data1, data2, data3);
         horizontal = leftPos[0];
         vertical = leftPos[1];
     }
